kafka-structured/services/ml-inference/requirements.txt
kafka-python>=2.0.2
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
